+++ src/handler.rs
@@ -1,8 +1,9 @@
 #[cfg(feature = "activity")]
 use crate::CacheEntry;
 use crate::commands::{change_reaction_role, track_emoji_usage, track_song};
+use crate::config::ReplyKind;
 use crate::constants::HTTP_CLIENT;
-use crate::{Data, ReplyKind, UserError};
+use crate::{Data, UserError};
 use anyhow::anyhow;
 use itertools::Itertools;
 use poise::serenity_prelude::json::json;
